- added date of birth in usermodel - added beneficiary columns to all benefits entities - change instalment column to decimal - added home loan benefid gui

+++ ZfssUZ/Models/Benefit/HomeLoanBenefit/HomeLoanBenefitModel.cs
@@ -17,7 +17,7 @@ namespace ZfssUZ.Models.Benefit.HomeLoanBenefit
         public long LoanCost { get; set; }
 
         [Required]
-        public int Instalment { get; set; }
+        public decimal Instalment { get; set; }
 
         [Required]
         public int Months { get; set; }
